# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Yapaia Go Core API",
    "version": "0.0.1",
    "description": "Automatisch generiert aus den tatsächlich registrierten Fastify-Routen (apps/core/src/index.ts, onRouteHook) und den @yapaia/shared-JSON-Schemas (packages/shared/src/schemas/). Quelle der Wahrheit ist der Code, nicht dieses Dokument -- bei Verdacht auf Abweichung: `pnpm openapi:check` (CI-Gate \"Spec aktuell\") bzw. docs/03-api-spec.md für den erzählenden Überblick. WebSocket (/ws/v1) und MQTT sind hier NICHT modelliert (OpenAPI eignet sich dafür nicht) -- siehe docs/03-api-spec.md §3/§4."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Core-API (relative Pfade -- funktioniert unverändert hinter HA-Ingress, W-15)"
    }
  ],
...
  "paths": {
...
    "/api/v1/routes/{id}": {
      "get": {
        "summary": "Route abrufen (Cache, TTL 1 h)",
        "tags": [
          "Routing"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Route"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/addons": {
      "get": {
        "summary": "Installierte Add-ons + Status",
        "tags": [
          "Add-ons"
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AddonManifest"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/addons/install": {
      "post": {
        "summary": "Installation starten (`{registry_id | url}`, Job)",
        "tags": [
          "Add-ons"
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort (Schema noch nicht kuratiert -- siehe apps/core/src/openapi/paths.ts)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/addons/install/{pendingId}/confirm": {
      "post": {
        "summary": "Berechtigungs-Bestätigung nach Prüfung abschließen",
        "tags": [
          "Add-ons"
        ],
        "parameters": [
          {
            "name": "pendingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort (Schema noch nicht kuratiert -- siehe apps/core/src/openapi/paths.ts)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/addons/{id}/enable": {
      "post": {
        "summary": "Add-on aktivieren",
        "tags": [
          "Add-ons"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort (Schema noch nicht kuratiert -- siehe apps/core/src/openapi/paths.ts)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/addons/{id}/disable": {
      "post": {
        "summary": "Add-on deaktivieren (Kill-Switch, W-10/W-11)",
        "tags": [
          "Add-ons"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort (Schema noch nicht kuratiert -- siehe apps/core/src/openapi/paths.ts)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/addons/{id}/mqtt/enable": {
      "post": {
        "summary": "MQTT-Discovery für Add-on aktivieren",
        "tags": [
          "Add-ons"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort (Schema noch nicht kuratiert -- siehe apps/core/src/openapi/paths.ts)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/addons/{id}/mqtt/disable": {
      "post": {
        "summary": "MQTT-Discovery für Add-on deaktivieren",
        "tags": [
          "Add-ons"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort (Schema noch nicht kuratiert -- siehe apps/core/src/openapi/paths.ts)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/addons/{id}": {
      "delete": {
        "summary": "Add-on deinstallieren (rückstandsfrei)",
        "tags": [
          "Add-ons"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort (Schema noch nicht kuratiert -- siehe apps/core/src/openapi/paths.ts)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Route": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Route identifier"
          },
          "distance_m": {
            "type": "number",
            "minimum": 0,
            "description": "Total route distance in meters"
          },
          "duration_s": {
            "type": "number",
            "minimum": 0,
            "description": "Total route duration in seconds (Valhalla time, calibrated)"
          },
          "geometry": {
            "type": "string",
            "description": "Encoded polyline (polyline6 format)"
          },
          "legs": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": {
                  "type": "integer",
                  "minimum": 0,
                  "description": "Leg index"
                },
                "distance_m": {
                  "type": "number",
                  "minimum": 0,
                  "description": "Leg distance in meters"
                },
                "duration_s": {
                  "type": "number",
                  "minimum": 0,
                  "description": "Leg duration in seconds"
                }
              },
              "required": [
                "index",
                "distance_m",
                "duration_s"
              ],
              "additionalProperties": false
            },
            "description": "Route segments between waypoints"
          },
          "maneuvers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": {
                  "type": "integer",
                  "minimum": 0,
                  "description": "Maneuver index in route"
                },
                "type": {
                  "type": "string",
                  "description": "Maneuver type (Valhalla enum, e.g. \"turn_left\", \"roundabout_enter\")"
                },
                "instruction": {
                  "type": "string",
                  "description": "Localized instruction text"
                },
                "street_names": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "Names of streets involved"
                },
                "distance_m": {
                  "type": "number",
                  "minimum": 0,
                  "description": "Length of maneuver segment in meters"
                },
                "begin_shape_index": {
                  "type": "integer",
                  "minimum": 0,
                  "description": "Start index in route geometry"
                },
                "lanes": {
                  "type": [
                    "array",
                    "null"
                  ],
                  "items": {
                    "type": "object",
                    "properties": {
                      "lane_index": {
                        "type": "integer",
                        "minimum": 0,
                        "description": "Lane index from left"
                      },
                      "is_usable": {
                        "type": "boolean",
                        "description": "Whether vehicle can use this lane"
                      },
                      "direction": {
                        "type": [
                          "string",
                          "null"
                        ],
                        "description": "Direction indicator if available"
                      }
                    },
                    "required": [
                      "lane_index",
                      "is_usable"
                    ],
                    "additionalProperties": false
                  },
                  "description": "Optional lane information"
                },
                "duration_s": {
                  "type": "number",
                  "minimum": 0,
                  "description": "Planned duration of this maneuver segment in seconds (Valhalla time), optional (E04-T2)"
                }
              },
              "required": [
                "index",
                "type",
                "instruction",
                "street_names",
                "distance_m",
                "begin_shape_index"
              ],
              "additionalProperties": false
            },
            "description": "Turn-by-turn instructions"
          },
          "speed_limits": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "begin_shape_index": {
                  "type": "integer",
                  "minimum": 0,
                  "description": "Start index in route geometry"
                },
                "end_shape_index": {
                  "type": "integer",
                  "minimum": 0,
                  "description": "End index in route geometry"
                },
                "kmh": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "minimum": 5,
                  "maximum": 130,
                  "description": "Speed limit in km/h (5–130 or null for unknown)"
                }
              },
              "required": [
                "begin_shape_index",
                "end_shape_index",
                "kmh"
              ],
              "additionalProperties": false
            },
            "description": "Speed restrictions along route"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "description": "Warning code identifier"
                },
                "message": {
                  "type": "string",
                  "description": "Human-readable warning message"
                }
              },
              "required": [
                "code",
                "message"
              ],
              "additionalProperties": false
            },
            "description": "Route warnings (e.g. \"restriction data missing\")"
          }
        },
        "required": [
          "id",
          "distance_m",
          "duration_s",
          "geometry",
          "legs",
          "maneuvers",
          "speed_limits",
          "warnings"
        ],
        "additionalProperties": false
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code identifier"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": [
                  "object",
                  "null"
                ],
                "description": "Optional additional error context"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false,
            "description": "Error object with code, message, and optional details"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "AddonManifest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-z0-9]([a-z0-9-]*[a-z0-9])?(\\.[a-z0-9]([a-z0-9-]*[a-z0-9])?)*$",
            "description": "Reverse-DNS-ish identifier; becomes the data/addons/{id} directory name"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "version": {
            "type": "string",
            "minLength": 1,
            "description": "Exact semver, e.g. \"1.2.0\""
          },
          "core_api": {
            "type": "string",
            "minLength": 1,
            "description": "Semver range, e.g. \"^1.0\""
          },
          "author": {
            "type": "string",
            "minLength": 1
          },
          "license": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "requires_online": {
            "type": "boolean"
          },
          "ui": {
            "type": "object",
            "properties": {
              "entry": {
                "type": "string",
                "minLength": 1
              },
              "widgets": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "minLength": 1
                    },
                    "name": {
                      "type": "string",
                      "minLength": 1
                    },
                    "slots": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "minLength": 1
                      }
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "slots"
                  ],
                  "additionalProperties": false
                }
              },
              "map_layers": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "minLength": 1
                    },
                    "name": {
                      "type": "string",
                      "minLength": 1
                    },
                    "source": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "source"
                  ],
                  "additionalProperties": false
                }
              },
              "settings_page": {
                "type": "boolean"
              }
            },
            "required": [
              "entry"
            ],
            "additionalProperties": false
          },
          "service": {
            "type": "object",
            "properties": {
              "runtime": {
                "type": "string",
                "enum": [
                  "node18",
                  "node20",
                  "external"
                ]
              },
              "entry": {
                "type": "string",
                "minLength": 1
              },
              "max_rss_mb": {
                "type": "number",
                "minimum": 16,
                "maximum": 4096
              }
            },
            "required": [
              "runtime",
              "entry"
            ],
            "additionalProperties": false
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "anyOf": [
                {
                  "enum": [
                    "pos.read",
                    "nav.read",
                    "nav.control",
                    "route.read",
                    "route.propose",
                    "map.layer.write",
                    "widget.register",
                    "events.publish",
                    "storage.own",
                    "ha.notify",
                    "camera.view"
                  ]
                },
                {
                  "pattern": "^net\\.fetch:.+$"
                }
              ]
            }
          }
        },
        "required": [
          "id",
          "name",
          "version",
          "core_api",
          "author",
          "license",
          "description",
          "permissions"
        ],
        "additionalProperties": false
      }
    }
  }
}